requirements.txt
openai
langchain_nvidia_ai_endpoints
langchain_community
faiss-cpu
python-dotenv
streamlit
pypdf
"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"